move and rename client fetch hook file
