+++ frontend/src/components/search/SaveSearchSelector.tsx
@@ -1,9 +1,9 @@
 import clsx from "clsx";
 import { noop } from "lodash";
+import { useClientFetch } from "src/hooks/useClientFetch";
 import { usePrevious } from "src/hooks/usePrevious";
 import { useSearchParamUpdater } from "src/hooks/useSearchParamUpdater";
 import { useUser } from "src/services/auth/useUser";
-import { useClientFetch } from "src/services/fetch/clientFetch";
 import { QueryContext } from "src/services/search/QueryProvider";
 import { SavedSearchRecord } from "src/types/search/searchRequestTypes";
 import { searchToQueryParams } from "src/utils/search/searchFormatUtils";
